Address review: require total DFA, move models to domain, use SymPy for matrix power
- Move DFA and DFATransition from jacobian.contracts to the regular_languages domain
- Require a transition for every (state, symbol) pair so the DFA is total
- Replace handwritten matrix-arithmetic kernel with SymPy Matrix powering
- Update tests for the new totality requirement and add a non-total DFA test
